pyproject.toml
[build-system]
requires = ["hatchling >= 1.21.0"]
build-backend = "hatchling.build"

[project]
name = "rubik-algorithm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mkdocs-macros-plugin>=1.3.9",
    "mkdocs-material>=9.6.18",
    "mkdocs-site-urls>=0.3.1",
    "numpy>=2.3.2",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.12",
]

[tool.ruff]
line-length = 100
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "PT", "UP"]

[tool.mypy]
strict = true
pretty = true
